Implementata funzione my posts

+++ laravel-one-to-many/app/Http/Controllers/Admin/PostController.php
@@ -90,6 +90,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
         return view('admin.posts.edit', compact('post'));
     }
 
@@ -102,6 +103,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         $request->validate($this->getValidators($post));
 
         $post->update($request->all());
@@ -117,8 +121,21 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::user()->id !== $post->user_id) abort(403);
+
         $post->delete();
 
-        return redirect()->route('admin.posts.index');
+        if (url()->previous() === route('admin.posts.edit', $post->slug)) {
+            return redirect()->route('admin.home')->with('status', "Post $post->title deleted");;
+        }
+        return redirect(url()->previous())->with('status', "Post $post->title deleted");;;
+    }
+
+
+    public function myindex() {
+
+        $posts = Post::where('user_id', Auth::user()->id)->paginate(30);
+
+        return view('admin.posts.index', compact('posts'));
     }
 }
